Refactor and enhance various components for improved functionality and consistency
- Updated ChangePassword.razor to improve form handling and validation response.
- Modified EmailConfirmation.razor to streamline email confirmation logic.
- Changed parameter types from int? to long? in Brand, Category, Color, and Size detail pages for consistency.
- Updated AccountsController to correct email confirmation and password reset URLs.
- Adjusted BrandsController, CategoriesController, ColorsController, and SizesController to use long for ID parameters.
- Refactored Storefront Controllers to replace int with long for various ID parameters.
- Enhanced SmtpService to include validation for email addresses and SMTP host configuration.
- Updated Order and Product related request and view models to use long for ID properties.
- Improved JavaScript in ProductSearchResults.cshtml for better handling of refinement groups and IDs.

+++ eCommerce.Backoffice.Server/Controllers/BrandsController.cs
@@ -34,7 +34,7 @@ namespace eCommerce.Backoffice.Server.Controllers
 
         [HttpGet("{id}")]
         [ResponseCache(Location = ResponseCacheLocation.None, NoStore = true)]
-        public async Task<ActionResult<BrandDto>> GetBrand(int id)
+        public async Task<ActionResult<BrandDto>> GetBrand(long id)
         {
             var brand = await _brandService.GetAsync(id);
 
@@ -72,7 +72,7 @@ namespace eCommerce.Backoffice.Server.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateBrand(int id, BrandDto brand)
+        public async Task<IActionResult> UpdateBrand(long id, BrandDto brand)
         {
             if (id != brand.Id)
             {
@@ -100,7 +100,7 @@ namespace eCommerce.Backoffice.Server.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteBrand(int id)
+        public async Task<IActionResult> DeleteBrand(long id)
         {
             try
             {
