+++ eCommerce.Storefront.Services/Messaging/OrderService/CreateOrderRequest.cs
@@ -4,7 +4,7 @@ namespace eCommerce.Storefront.Services.Messaging.OrderService
 {
     public class CreateOrderRequest
     {
-        public int DeliveryId { get; set; }
+        public long DeliveryId { get; set; }
         public Guid BasketId { get; set; }
         public string CustomerEmail { get; set; }
     }
