+++ eCommerce.Storefront.Services/Messaging/ProductCatalogService/ModifyBasketRequest.cs
@@ -15,7 +15,7 @@ namespace eCommerce.Storefront.Services.Messaging.ProductCatalogService
         public Guid BasketId { get; set; }
         public IList<long> ItemsToRemove { get; set; }
         public IList<ProductQtyUpdateRequest> ItemsToUpdate { get; set; }
-        public int SetShippingServiceIdTo { get; set; }
+        public long SetShippingServiceIdTo { get; set; }
         public IList<long> ProductsToAdd { get; set; }
     }
 }
