+++ eCommerce.Storefront.Services/ViewModels/BasketItemView.cs
@@ -2,11 +2,11 @@ namespace eCommerce.Storefront.Services.ViewModels
 {
     public class BasketItemView
     {
-        public int Id { get; set; }
-        public int ProductId { get; set; }
+        public long Id { get; set; }
+        public long ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductSizeName { get; set; }
-        public int ProductTitleId { get; set; }
+        public long ProductTitleId { get; set; }
         public int Qty { get; set; }
         public string ProductPrice { get; set; }
         public string LineTotal { get; set; }
